# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CharacterResponse:
    properties:
      character:
        description: Display name of the character
        example: Ushiromiya Battler
        type: string
      characterId:
        description: Numeric character identifier
        example: "10"
        type: string
      limit:
        description: Maximum results per page
        example: 50
        type: integer
      offset:
        description: Pagination offset
        example: 0
        type: integer
      quotes:
        description: Paginated list of quotes
        items:
          type: object
        type: array
      total:
        description: Total number of matching quotes
        example: 4304
        type: integer
    type: object
  dto.ContextResponse:
    properties:
      after:
        description: Dialogue lines after the target quote
        items:
          type: object
        type: array
      before:
        description: Dialogue lines before the target quote
        items:
          type: object
        type: array
      quote:
        description: The target quote
        type: object
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        description: Error message
        example: query parameter 'q' is required
        type: string
    type: object
  dto.HigurashiStatsResult:
    properties:
      characterNames:
        additionalProperties:
          type: string
        description: Map of character ID to display name
        type: object
      interactionCounts:
        additionalProperties:
          type: integer
        description: Map of character pair key to interaction count
        type: object
      interactions:
        description: Top character interaction pairs
        items:
          $ref: '#/definitions/dto.InteractionPair'
        type: array
      linesPerArc:
        additionalProperties:
          additionalProperties:
            type: integer
          type: object
        description: Dialogue line counts per arc per character
        type: object
      topSpeakers:
        description: Top speakers by dialogue line count
        items:
          $ref: '#/definitions/dto.SpeakerStat'
        type: array
    type: object
  dto.InteractionPair:
    properties:
      charA:
        description: First character ID
        example: "10"
        type: string
      charB:
        description: Second character ID
        example: "27"
        type: string
      count:
        description: Number of shared dialogue exchanges
        example: 1701
        type: integer
      nameA:
        description: First character name
        example: Ushiromiya Battler
        type: string
      nameB:
        description: Second character name
        example: Beatrice
        type: string
    type: object
  dto.SearchAPIResponse:
    properties:
      limit:
        description: Maximum results per page
        example: 30
        type: integer
      offset:
        description: Pagination offset
        example: 0
        type: integer
      query:
        description: The search query that was used
        example: without love
        type: string
      results:
        description: Matching quotes with relevance scores
        items:
          $ref: '#/definitions/dto.SearchResult'
        type: array
      total:
        description: Total number of matches
        example: 25
        type: integer
    type: object
  dto.SearchResult:
    properties:
      quote:
        description: The matched quote
        type: object
      score:
        description: Relevance score (100 = exact match)
        example: 100
        type: integer
    type: object
  dto.SoundEffect:
    properties:
      afterClip:
        example: 0
        type: integer
      filename:
        example: umise_047
        type: string
    type: object
  dto.SpeakerStat:
    properties:
      characterId:
        description: Numeric character identifier
        example: "10"
        type: string
      count:
        description: Number of dialogue lines
        example: 4304
        type: integer
      name:
        description: Display name of the character
        example: Ushiromiya Battler
        type: string
    type: object
...
  umineko_quote_internal_dto.HigurashiQuote:
    properties:
      arc:
        example: onikakushi
        type: string
      audioCharMap:
        additionalProperties:
          type: string
        example:
          "30101088": "10"
          "30101089": "10"
        type: object
      audioId:
        example: 30101088, 30101089
        type: string
      audioTextMap:
        additionalProperties:
          type: string
        example:
          "30101088": You're insane!!
          "30101089": Stop it already
        type: object
      character:
        example: Ushiromiya Battler
        type: string
      characterId:
        example: "10"
        type: string
      contentType:
        example: ""
        type: string
      episode:
        example: 3
        type: integer
      index:
        description: Position index in the parsed script (stable across text changes)
        type: integer
      soundEffects:
        items:
          $ref: '#/definitions/dto.SoundEffect'
        type: array
      text:
        example: You're insane!! Stop it already, damn iiiiiiiiit!!
        type: string
      textHtml:
        example: You&#39;re insane!! Stop it already, damn iiiiiiiiit!!
        type: string
      textJp:
        type: string
      textJpHtml:
        type: string
    type: object
...
info:
  contact:
    email: FAA@auaurora.moe
    name: Featherine Augustus Aurora
    url: https://x.com/FeatherineFAA
  description: API for searching and browsing quotes from Umineko no Naku Koro ni
    and Higurashi no Naku Koro ni
  title: When They Cry Quote API
  version: "1.0"
paths:
  /higurashi/browse:
    get:
      description: Browse all Higurashi quotes with optional filters and pagination
      parameters:
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      - description: Filter by character ID
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: character
        type: string
      - description: 'Interaction filter: first character ID (requires interactionB)'
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: interactionA
        type: string
      - description: 'Interaction filter: second character ID (requires interactionA)'
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: interactionB
        type: string
      - default: 50
        description: Maximum results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by arc number
        in: query
        name: episode
        type: integer
      - description: Filter by arc name (e.g. onikakushi, watanagashi)
        in: query
        name: arc
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CharacterResponse'
      summary: Browse quotes
      tags:
      - higurashi
...
  /higurashi/context/{audioId}:
    get:
      description: Returns surrounding dialogue lines for a specific Higurashi quote
      parameters:
      - description: Audio ID of the quote (e.g. ps3/s01/01/hrs010010)
        in: path
        name: audioId
        required: true
        type: string
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      - default: 5
        description: Number of context lines before and after
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get quote context
      tags:
      - higurashi
  /higurashi/nearest-voiced/{audioId}:
    get:
      description: Returns the audio ID of the nearest Higurashi quote with voice
        audio in a given direction
      parameters:
      - description: Audio ID of the reference quote (e.g. ps3/s01/01/hrs010010)
        in: path
        name: audioId
        required: true
        type: string
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      - default: next
        description: Direction to search
        enum:
        - next
        - prev
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Find nearest voiced quote
      tags:
      - higurashi
  /higurashi/quote/{audioId}:
    get:
      description: Returns a specific Higurashi quote identified by its audio ID (may
        contain slashes)
      parameters:
      - description: Audio ID of the quote (e.g. ps3/s01/01/hrs010010)
        in: path
        name: audioId
        required: true
        type: string
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/umineko_quote_internal_dto.HigurashiQuote'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get quote by audio ID
      tags:
      - higurashi
  /higurashi/quote/index/{index}:
    get:
      description: Returns a specific Higurashi quote identified by its position index
        in the parsed script
      parameters:
      - description: Index of the quote in the parsed script
        in: path
        name: index
        required: true
        type: integer
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/umineko_quote_internal_dto.HigurashiQuote'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get quote by index
      tags:
      - higurashi
  /higurashi/random:
    get:
      description: Returns a random Higurashi quote with optional filters
      parameters:
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      - description: Filter by character ID
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: character
        type: string
      - description: Filter by arc number
        in: query
        name: episode
        type: integer
      - description: Filter by arc name (e.g. onikakushi, watanagashi)
        in: query
        name: arc
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/umineko_quote_internal_dto.HigurashiQuote'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get random quote
      tags:
      - higurashi
  /higurashi/search:
    get:
      description: Search for Higurashi quotes by text query with optional filters
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: en
        description: Language
        in: query
        name: lang
        type: string
      - default: 30
        description: Maximum results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by character ID
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: character
        type: string
      - description: 'Interaction filter: first character ID (requires interactionB)'
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: interactionA
        type: string
      - description: 'Interaction filter: second character ID (requires interactionA)'
        enum:
        - narrator
        - misc_voices
        - keiichi
        - rena
        - mion
        - satoko
        - rika
        - shion
        - satoshi
        - tomitake
        - takano
        - irie
        - ooishi
        - hanyuu
        - akasaka
        - okonogi
        - kasai
        - kimiyoshi
        - oryou
        - teppei
        - rina
        - chie
        - tomoe
        - madoka
        - yamaoki
        - fujita
        - natsumi
        - chisato
        - tamako
        - akira
        - miyuki
        - otobe
        - towada
        - riku
        - ouka
        - kumagai
        - nagisa
        - akane
        - arakawa
        - maeno
        in: query
        name: interactionB
        type: string
      - description: Filter by arc number
        in: query
        name: episode
        type: integer
      - description: Filter by arc name (e.g. onikakushi, watanagashi)
        in: query
        name: arc
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SearchAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Search quotes
      tags:
      - higurashi
  /higurashi/stats:
    get:
      description: Returns statistics about Higurashi quotes including top speakers
        and character interactions
      parameters:
      - description: Filter by arc number, 0 for all
        in: query
        name: episode
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HigurashiStatsResult'
      summary: Get quote statistics
      tags:
      - higurashi
...
schemes:
- https
- http
swagger: "2.0"